refactor(template_sia): move loop orchestration into src/, drop scripts shim
src/loop.py now holds the real harness orchestration (build_run_config,
run_sia_loop_project) instead of re-exporting scripts/sia_loop_impl.py via a
sys.path injection. scripts/run_sia_loop.py is a thin caller. The sanctioned
infrastructure.sia import is declared in manuscript/layer_contract.yaml
(allow_infrastructure_imports: [src/loop.py]), matching sibling exemplar
template_autoresearch_project. Restores the thin-orchestrator contract every
other exemplar follows.

- delete scripts/sia_loop_impl.py + its sys.path shim
- update AGENTS.md, scripts/AGENTS.md, src/AGENTS.md, docs/architecture.md
- tests unchanged (all import from src.loop); coverage 99.69%, strict drift green
