+++ projects/templates/template_sia/docs/architecture.md
@@ -24,13 +24,12 @@ flowchart TB
     REF[reference/]
   end
 
-  SCR[scripts/run_sia_loop.py] --> IMPL[scripts/sia_loop_impl.py]
-  IMPL --> LR
-  LOOP -.re-export shim.-> IMPL
+  SCR[scripts/run_sia_loop.py] --> LOOP
+  LOOP --> LR
   LR --> TL
   LR --> ER
   LR --> CL
-  IMPL --> TASK
+  LOOP --> TASK
   LR -->|live=False| FIX[src/fixtures/]
   LR -->|live=True| LLM[infrastructure/llm]
 ```
@@ -39,7 +38,7 @@ flowchart TB
 | --- | --- | --- |
 | Layer 1 | `infrastructure/sia/` | Task validation, evaluation subprocess, context ledger, generation state machine |
 | Layer 2 | `projects/templates/template_sia/src/` | Project config, fixture paths, manuscript variables, task-specific reference agent |
-| Orchestration | `scripts/` | CLI flags (`--live-sia`), stdout paths for manifest collection; `sia_loop_impl.py` holds the actual `build_run_config()`/`run_sia_loop_project()` orchestration, re-exported by `src/loop.py` as a compatibility shim |
+| Orchestration | `scripts/` | CLI flags (`--live-sia`), stdout paths for manifest collection; `run_sia_loop.py` is a thin caller that imports `build_run_config()`/`run_sia_loop_project()` from `src/loop.py`, where the orchestration lives |
 
 ## Generation artifact tree
